Added new animations to the chameleon, walking, sitting and idle is going to be worked on shortly
Updated chameleon texture and is WIP

Made the chameleon and its animation faster because it was hella slow.

I HAVE ALSO MADE IT TAMAMBE WITH APPLES AND IT FOLLOWS YOU AND YOU CAN MAKE IT SIT.

Last thing is that ive also made it spawn in the world in the jungles!